requirements.txt
tiktoken
openai
python-dotenv
azure-identity
langchain
langchain-openai
azure-identity